% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomsearchResult.R
\name{summary.RandomsearchResult}
\alias{summary.RandomsearchResult}
\title{Generate result summary}
\usage{
\method{summary}{RandomsearchResult}(object, ...)
}
\arguments{
\item{object}{[\code{\link[ParamHelpers]{OptPath}}]\cr
Optimization path.}

\item{...}{[any] \cr
Currently ignored.}
}
\description{
Generate results from randomsearch run.
For normal single objective runs the best result is returned.
For mulit-objective runs the pareto front is returned.
}
